move a few files around for a more consistent structure
