+++ src/Core/CMakeLists.txt
@@ -1,6 +1,9 @@
 
 
 add_subdirectory(QueryParser)
+add_library(PlaceVisitor PlaceVisitor.cpp)
+target_link_libraries(PlaceVisitor QueryParser)
+
 add_subdirectory(TAPN)
 add_subdirectory(TAPNParser)
