+++ src/Core/TAPN/CMakeLists.txt
@@ -2,6 +2,4 @@
 
 
 add_library(TAPN InhibitorArc.cpp TimedArcPetriNet.cpp TimedPlace.cpp TimeInterval.cpp TransportArc.cpp OutputArc.cpp TimedInputArc.cpp TimedTransition.cpp TimeInvariant.cpp)
-
-
-target_link_libraries(TAPN DiscreteVerification)
+target_link_libraries(TAPN PlaceVisitor)
