+++ src/DiscreteVerification/CMakeLists.txt
@@ -3,13 +3,10 @@ add_subdirectory(DataStructures)
 add_subdirectory(Util)
 add_subdirectory(SearchStrategies)
 add_subdirectory(VerificationTypes)
+add_subdirectory(Generators)
 
 
 add_library(DiscreteVerification DeadlockVisitor.cpp
-        DiscreteVerification.cpp
-        Generator.cpp
-        PlaceVisitor.cpp
-        ReducingGenerator.cpp
-        InterestingVisitor.cpp)
+        DiscreteVerification.cpp)
 
-target_link_libraries(DiscreteVerification DataStructures Util SearchStrategies VerificationTypes)
+target_link_libraries(DiscreteVerification DataStructures Util SearchStrategies VerificationTypes Generators)
