+++ src/DiscreteVerification/Generators/CMakeLists.txt
@@ -1,0 +1,4 @@
+
+add_library(Generators Generator.cpp ReducingGenerator.cpp InterestingVisitor.cpp)
+
+
